Pass only connector id to workflow
